+++ src/RestApi/Emulator/EmulatorManagement.ts
@@ -3,11 +3,13 @@ import IOptions from "../IOptions";
 import IEmulator, { IEmulatorCreatable } from './IEmulator';
 import Emulator from './Emulator';
 import { IEmulatorFilter } from './IEmulatorFilter';
+import DeviceManagement from '../Device/DeviceManagement';
+import IDevice from '../Device/IDevice';
 
 export default class EmulatorManagement {
 	public static readonly RESOURCE: string = 'emulator';
 
-	constructor(private options: IOptions) {}
+	constructor(private options: IOptions, private deviceManagement: DeviceManagement) {}
 
 	public async list(filter?: IEmulatorFilter) {
 		const response = await getResource(this.options, EmulatorManagement.RESOURCE, filter);
@@ -15,8 +17,16 @@ export default class EmulatorManagement {
 		return data.map((item: IEmulator) => new Emulator(item));
 	}
 
-	public async create(settings: IEmulatorCreatable): Promise<void> {
-		await postResource(this.options, EmulatorManagement.RESOURCE, JSON.stringify(settings));
+	public async create(settings: IEmulatorCreatable): Promise<IDevice> {
+		const { headers } = await postResource(this.options, EmulatorManagement.RESOURCE, JSON.stringify(settings));
+		const headerLocation = headers.get('location');
+
+		if (!headerLocation) {
+			throw new Error(`Api didn't return location header to created ${EmulatorManagement.RESOURCE}.`);
+		}
+
+		const deviceUid = headerLocation.split('/').slice(-1)[0];
+		return await this.deviceManagement.get(deviceUid);
 	}
 
 	public async delete(deviceUid: string): Promise<void> {