+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Optional field `autoEnableTimeoutMs` when disabling device peer recovery
 - Device auto recovery
 - Update types for bulk operations
+- Device is returned after Emulator creation
 
 ## [1.14.0] - 2022-08-03
 ### Added